only authors of campgrounds can edit or delete campgrounds- client and server side- created middleware isLoggedIn and isAuthor
