+++ routes/reviews.js
@@ -7,17 +7,9 @@ const Review = require('../models/review');
 const catchAsync = require('../utils/catchAsync');
 const ExpressError = require('../utils/ExpressError');
 
-const { reviewSchema } = require('../schemas.js');
+// const { reviewSchema } = require('../schemas.js');
+const { validateReview } = require('../middleware');
 
-const validateReview = (req, res, next) => {
-    const { error } = reviewSchema.validate(req.body);
-    if (error) {
-        const msg = error.details.map(el => el.message).join(',')
-        throw new ExpressError(msg, 400)
-    } else {
-        next();
-    }
-}
 
 router.post('/', validateReview, catchAsync(async (req, res) => {
     const campground = await Campground.findById(req.params.id);
