+++ views/campgrounds/show.ejs
@@ -23,16 +23,18 @@
                         €<%= campground.price %>/night
                     </li>
                 </ul>
-                <div class="card-body">
-                    <a href="/campgrounds/<%=campground._id%>/edit" class="card-link btn btn-warning">Edit
-                        Campground</a>
-                    <form class="d-inline" action="/campgrounds/<%=campground._id %>?_method=DELETE" method="POST">
-                        <button class="card-link btn btn-danger">Delete</button>
-                    </form>
-                </div>
-                <div class="card-footer text-muted">
-                    <a href="/campgrounds" class="card-link">All Campgrounds</a>
-                </div>
+                <% if (currentUser && campground.author.equals(currentUser)) { %>
+                    <div class="card-body">
+                        <a href="/campgrounds/<%=campground._id%>/edit" class="card-link btn btn-warning">Edit
+                            Campground</a>
+                        <form class="d-inline" action="/campgrounds/<%=campground._id %>?_method=DELETE" method="POST">
+                            <button class="card-link btn btn-danger">Delete</button>
+                        </form>
+                        <% } %>
+                    </div>
+                    <div class="card-footer text-muted">
+                        <a href="/campgrounds" class="card-link">All Campgrounds</a>
+                    </div>
             </div>
 
         </div>
